server/requirements.txt
fastapi>=0.115
uvicorn>=0.30
neo4j>=5.0
pydantic>=2.0
python-dotenv>=1.0
openai>=1.0
pywinrm>=0.4.3
paramiko>=3.0